+++ packages/app/jest.setenv.vars.js
@@ -1,0 +1,1 @@
+process.env.NEXT_PUBLIC_API_URL = "https://egapro-preprod.dev.fabrique.social.gouv.fr/api";